Issue 193:
Initial commit. Added code to set the end time of created hangouts to 24 hours later (given that the
user does not choose an end time for themselves). Changed the display format so that these day long (or
greater) hangout sessions show up as an end time of "To Be Announced".

Any UI that was created or changed is a temporary placeholder. I suck at UI. Help would be greatly
appreciated