issue-193 Part 2: End hangout button
Thanks to some help, I was able to get the create the end hangout button.
The button itself is a little long in width and I do not know how to correct that...

The end hangout button function does work, but it is a hack. I tried to use the
edit hangout Meteor call and, in doing so, I found hacky ways to obtain all
the information needed for that call.

Another issue is that UI does not immediately update. On my computer the
new end time was slow to appear. Also, if you hit the end hangout button
multiple times, you will create new events. That seems like a bug in the
edit hanogut meteor call, but I am not sure