Release 0.9.90 (#754)
* Issue 691 hangout participant count (#719)

* fix for bug 690: admins should be able to access 24/7 hangout (#693) (#694)

* Issue-691. Show number of participants in 24/7 hangout




Issue-691. show number of hangout participants in all study group and my study group pages


issue-691.  Only decrement participant count when the logged in user joins the hangout room


Issue-691 Decrement participant count when value is positive

* Remove unique index comment

* fix 'participants_count' of undefined console err

* Get rid of session variable

* Move server-side logic to new files

* Enable seed data

* minor improvements

* Show laptop icon when people join 24/7 hangout

* Fix account deletion. (#745)

* fix typo

* Schedule hangout default studygroup name (#743)

* Default study group name 

when schedule a hangout in single study group
duplicate hangout in single study group

* Default study group name 

value is not displayed in dropdown input area unless trigger('change') is fired

* Fixed error

* Issue 667 create slack page (#744)

* Create new Slack page

* Add helpful links

* Minor content edits!

Note: I removed the slackin plugin in favor of a direct link because I noticed some positioning issues with the form that appears after a user clicks on the plugin; it's sometimes out of reach, and will make the user think that nothing happened.

* Show study group name in disabled input box in modal (#748)

* Refactoring 1 (#752)

* fix for bug 690: admins should be able to access 24/7 hangout (#693) (#694)

* refactor (routes) remove subs from router level

* refactor (hangout_logged_out) make use of new syntax

* refactor (hangouts) - remove dead code, restructure single hangout view