Remove the loading screen when loading more
When the Load More button is pressed, the hangouts disappear and
the loading state momentarily flashes up. This was caused by
`subscriptionsReady` being false when getting more Hangouts.

I also moved the current and past headers into the `if` conditionals to
only display if there are any current or past hangouts to display

Closes #1003